Delete unused partial `products/product`
Its content is most like `products/index`, and a little like `carts/cart_item`, but there is not enough duplication to be worth extracting. And extracting it from `products/index` for naming purposes wouldn’t help either, since the embedded form is important to and specific to the index page, and I want it visible in that view.